% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.ipeb}
\alias{plot.ipeb}
\title{Plot the fitted iPEB marker weights}
\usage{
\method{plot}{ipeb}(x, ...)
}
\arguments{
\item{x}{A fitted \code{"ipeb"} object.}

\item{...}{Passed to \code{\link[graphics]{barplot}}.}
}
\value{
The fitted object, invisibly.
}
\description{
Plot the fitted iPEB marker weights
}
